Add files via upload
Assignment instructions ask for a Markdown header comment with assignment name, author, version, and semester information. I wasn't clear on what I was expected to input for version - I input our course title to demonstrate that I understand how to include a Markdown header and read the instructions closely, but wanted to note my confusion. If this is not satisfactory, please let me know what "version" is asking for and I will resubmit a corrected version. Thank you!